+++ tests/numpy/test_generic.py
@@ -55,10 +55,13 @@ class GenericTestCase(NumpyBaseTestCase):
 class NumpyProgressTestCase(NumpyBaseTestCase):
     def test_select_with_progress(self):
         progress = self.client.execute_with_progress('SELECT 2')
-        self.assertEqual(
-            list(progress),
-            [(1, 0), (1, 0)] if self.server_version > (20,) else [(1, 0)]
-        )
+        if self.server_version > (20,):
+            # total_rows in progress depends on server version
+            self.assertIn(
+                list(progress), ([(1, 0), (1, 0)], [(1, 1), (1, 1)])
+            )
+        else:
+            self.assertEqual(list(progress), [(1, 0)])
         self.assertEqual(progress.get_result(), [(2,)])
         self.assertTrue(self.client.connection.connected)